style(client appointments): appointment status dropdown and view details - Macatual
