+++ ServEase/src/app/styles/client-appointments.module.css
@@ -166,7 +166,8 @@
   background-color: #c9bbaf; /* Light brown background from image */
   border-radius: 50px;
   padding: 4px;
-  flex-shrink: 0; /* Prevents shrinking on smaller viewports */
+  flex-shrink: 0;
+  gap: 2px; /* Prevents shrinking on smaller viewports */
 }
 
 .filterButton {
@@ -416,4 +417,130 @@
   .logoText {
     display: none; /* Hides brand name text on very small screens to save space */
   }
+}
+
+/*dropdown*/
+.dropdownMenu {
+  position: absolute;
+  right: 0;
+  margin-top: 7rem;
+  width: 186px;
+  background-color: white;
+  border-radius: 12px;
+  overflow: hidden;
+  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
+  z-index: 100;
+}
+
+.dropdownItem {
+  padding: 15px 16px;
+  color: #241f1b;
+  text-decoration: none;
+  display: block;
+  transition: background 0.2s ease;
+  font-size: 18px;
+  cursor: pointer;
+}
+
+.dropdownItem:hover {
+  background-color: #a68465;
+  color: white;
+}
+
+.activeTop {
+  border-top-left-radius: 8px;
+  border-top-right-radius: 8px;
+}
+
+.activeBottom {
+  border-bottom-left-radius: 8px;
+  border-bottom-right-radius: 8px;
+}
+
+/*view details*/
+.modalOverlay {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100vw;
+  height: 100vh;
+  background-color: rgba(17, 17, 17, 0.5); /* semi-transparent dark */
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  z-index: 2000;
+}
+.calendarSelectChangeSize {
+  margin: 0 auto;
+  max-width: 500px;
+  width: 65vh;
+  position: relative;
+  box-shadow: 0px 16px 24px rgba(0, 0, 0, 0.06);
+  border-radius: 10px;
+  background-color: var(--color-white);
+  border: 1px solid rgba(0, 0, 0, 0.06);
+  box-sizing: border-box;
+  overflow: hidden;
+  text-align: left;
+  font-size: var(--font-size-16);
+  color: var(--color-dimgray);
+  font-family: var(--font-dm-sans);
+  min-height: 54vh;
+  padding: 0 auto;
+}
+.rowContainer {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  width: 100%;
+  margin-bottom: 3%;
+}
+
+.facilityName {
+  text-align: left;
+  flex: 1;
+}
+
+.facilityNameCap {
+  color: var(--color-tan);
+  text-align: right;
+  flex: 1;
+}
+.facilityNameParent {
+  position: relative;
+  margin: 0 10%;
+  margin-top: 8%;
+  width: 390px;
+  height: 191px;
+}
+.serviceNameParent {
+  position: relative;
+  margin: 0 auto;
+  width: 390px;
+  height: 157px;
+}
+.dividerIcon {
+  position: relative;
+  margin: 5% auto;
+  width: 390px;
+  height: 1px;
+  background-color: #dfceba;
+}
+.dividerIcon1 {
+  position: relative;
+  margin: -10% auto;
+  width: 390px;
+  height: 1px;
+  background-color: #dfceba;
+}
+.rowContainerTotal {
+  position: relative;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  width: 100%;
+  max-width: 390px;
+  margin: 0 auto;
+  margin-top: 15%;
+  margin-bottom: 8%;
 }
